GRADE - [FIX] GED average score's data type on DTO (int -> BigDecimal)

+++ info/src/main/java/kr/hs/entrydsm/husky/domain/application/dto/ScoreResponse.java
@@ -9,6 +9,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 @NoArgsConstructor
@@ -18,7 +20,7 @@ public class ScoreResponse {
 
     private GradeType gradeType;
 
-    private int gedAverageScore;
+    private BigDecimal gedAverageScore;
 
     private int volunteerTime;
     private int fullCutCount;
